[Feature] 나의 상품조회 API 구현

+++ src/main/java/com/github/secondproject/display/controller/DisplayController.java
@@ -37,7 +37,7 @@ public class DisplayController {
     @GetMapping("/{product_id}")
     public ResponseEntity<?> getProductDetail(
             @Parameter(name = "product_id", description = "상품 ID", example = "1")
-            @PathVariable("product_id") Integer productId ) {
+            @PathVariable("product_id") Long productId ) {
         ProductEntity response = displayService.getProduct(productId);
 
         return ResponseEntity.ok(response);
